+++ changelog.md
@@ -1,3 +1,6 @@
+# Version 2023.2.6 (2023-02-06)
+- Add missing bind_collector
+- 
 # Version 2023.2.5 (2023-02-05)
 - Fix GitHub build/publish
 - Add comments to parameter_visibility